Cambiar campo cantidad a entero - ALTER TABLE detalle_cotizaciones MODIFY cantidad INT - Input cantidad con step=1 y min=1 sin decimales - parseInt en cálculos y validaciones - Mostrar cantidad como entero en tabla de items

+++ database.sql
@@ -33,7 +33,7 @@ CREATE TABLE IF NOT EXISTS detalle_cotizaciones (
     id INT AUTO_INCREMENT PRIMARY KEY,
     cotizacion_id INT NOT NULL,
     producto_id INT NOT NULL,
-    cantidad DECIMAL(10,2) NOT NULL,
+    cantidad INT NOT NULL,
     precio_unitario DECIMAL(10,2) NOT NULL,
     total DECIMAL(10,2) NOT NULL,
     FOREIGN KEY (cotizacion_id) REFERENCES cotizaciones(id) ON DELETE CASCADE,
